backend/app/requirements.txt
# =========================
# FASTAPI BACKEND
# =========================
fastapi==0.115.0
uvicorn==0.30.6

# =========================
# DATABASE
# =========================
pymongo==4.8.0
python-dotenv==1.0.1

# =========================
# AUTHENTICATION
# =========================
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose==3.3.0
email-validator==2.1.1

# =========================
# AI / ML
# =========================
numpy==1.26.4
opencv-python==4.8.1.78
scikit-learn==1.5.1
tensorflow==2.15.0

# =========================
# UTILITIES
# =========================
pydantic==2.9.2